chore: extract address derivation from the Faucet module.
